+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## unreleased
+
+* Add document upload API
+
 ## 2.72.1
 
 * Add additional properties to `PaymentMethodNonce`